# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "version": "2.0.0",
    "title": "Slimmom API docs",
    "description": "API documentation for GoIT project 'Slimmom'. [Backend github repository](https://github.com/Dimitr-Remicio/SlimMomAPI)",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [{ "url": "https://slimmomsapi-dev-bbqt.3.us-1.fl0.io/api" }],
  "consumes": ["application/json"],
  "produces": ["application/json"],
...
  "paths": {
...
    "/days": {
      "patch": {
        "tags": ["Day"],
        "summary": "Post an eaten product",
        "security": [{ "Bearer": [] }],
        "parameters": [],
        "requestBody": {
          "description": "Object with date and eaten product's id",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchtDayRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatchDayResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request (invalid request body) / No token provided",
            "content": {}
          },
          "401": {
            "description": "Unauthorized (invalid access token)",
            "content": {}
          },
          "404": {
            "description": "Product not found / Habit not found / Invalid user / Invalid session",
            "content": {}
          }
        }
      },
      "delete": {
        "tags": ["Day"],
        "summary": "Delete eaten product",
        "security": [{ "Bearer": [] }],
        "parameters": [],
        "requestBody": {
          "description": "Object with eaten product's id you want to delete",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteDayRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteDayResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request (invalid request body) / No token provided",
            "content": {}
          },
          "401": {
            "description": "Unauthorized (invalid access token)",
            "content": {}
          },
          "404": {
            "description": "Product not found / Day not found / Invalid user / Invalid session",
            "content": {}
          }
        }
      }
    },
    "/days/info": {
      "post": {
        "tags": ["Day"],
        "summary": "Get info for day",
        "security": [{ "Bearer": [] }],
        "parameters": [],
        "requestBody": {
          "description": "Object with date of the day you want to get info for",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetDayRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation / No such day found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetDayResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request (invalid request body) / No token provided",
            "content": {}
          },
          "401": {
            "description": "Unauthorized (invalid access token)",
            "content": {}
          },
          "404": {
            "description": "Invalid user / Invalid session",
            "content": {}
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PatchtDayRequest": {
        "type": "object",
        "required": ["date", "productId", "weight"],
        "properties": {
          "date": {
            "type": "string",
            "description": "Day's date, use current date",
            "example": "2023-11-07"
          },
          "productId": {
            "type": "string",
            "description": "Product's id",
            "example": "5d51694802b2373622ff56d6"
          },
          "weight": {
            "type": "number",
            "description": "Product's weight",
            "example": 100,
            "minimum": 1,
            "maximum": 3000
          }
        }
      },
      "PatchDayResponse": {
        "type": "object",
        "properties": {
          "addedProduct": {
            "type": "object",
            "properties": {
              "product": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    
                    "_id": {
                      "type": "string",
                      "description": "Eaten product's id",
                      "example": "5d51694802b2373622ff56d6"
                    },
                    "title": {
                      "type": "string",
                      "description": "Eaten product's description",
                      "example": "Whole-ground oat flour Lifestyle"
                    },
                    "weight": {
                      "type": "number",
                      "description": "Eaten product's weight",
                      "example": 100
                    },
                    "calories": {
                      "type": "number",
                      "description": "kcal consumed",
                      "example": 368
                    },
                    "groupBloodNotAllowed": {
                      "type": "array",
                      "description": "Product's allowed/not allowed status on all blood types (first item is null - 0 index in array)",
                      "items": {
                         "oneOf": [{ "type": "null" }, { "type": "boolean" }]
                      },
                      "example": [null, true, false, false, false]
                   }
                  }
                }
              },
              "dayId": {
                "type": "string",
                "description": "Eaten product's id",
                "example": "654a5102b07d7d39528b3c84"
              },
              "UserId": {
                "type": "string",
                "description": "Eaten product's id",
                "example": "654a5102b07d7d39528b3c84"
              },
              "weight": {
                "type": "number",
                "description": "Eaten product's weight",
                "example": 100
              },
              "calories": {
                "type": "number",
                "description": "kcal consumed",
                "example": 368
              }
            }
          },
          "summary": {
            "type": "object",
            "description": "Updated summary (or newSummary)",
            "properties": {
              "id": {
                "type": "string",
                "description": "Day's id",
                "example": "654a481100b1c10dd00cd59d"
              },
              "date": {
                "type": "string",
                "description": "Day's date",
                "example": "2023-11-07"
              },
              "userId": {
                "type": "string",
                "description": "User's id",
                "example": "654a481000b1c10dd00cd59b"
              },
              "dailyRate": {
                "type": "number",
                "description": "Daily rate in kcal",
                "example": 1604
              },
              "kcalLeft": {
                "type": "number",
                "description": "Kcal left to consume",
                "example": 1036
              },
              "consumed": {
                "type": "number",
                "description": "Kcal consumed",
                "example": 568
              },
              
              "percentsOfDailyRate": {
                "type": "number",
                "description": "Percents consumed of daily rate",
                "example": 35.41147132169576
              },
              "createdAt": {
                "type": "string",
                "description": "Day summary's creation date",
                "example": "2023-11-07T14:22:09.374Z"
              },
              "updatedAt": {
                "type": "string",
                "description": "Day summary's update date",
                "example": "2023-11-07T15:12:35.539Z"
              }
              
            }
          }
        }
      },
      "DeleteDayRequest": {
        "type": "object",
        "required": ["dayId", "productId","sumId"],
        "properties": {
          "dayId": {
            "type": "string",
            "description": "Day's id",
            "example": "654a5102b07d7d39528b3c84"
          },
          "eatenProductId": {
            "type": "string",
            "description": " Product's id",
            "example": "5d51694802b2373622ff56d6"
          },
          "sumId": {
            "type": "string",
            "description": " Sumarys's id",
            "example": "654a481100b1c10dd00cd59d"
          }
        }
      },
      "DeleteDayResponse": {
        "type": "object",
        "description": "Updated summary",
        "properties": {
          "updateDay": {
            "type": "object",
            "description": "Updated summary (or newSummary)",
            "properties": {
              "id": {
                "type": "string",
                "description": "Day's id",
                "example": "654a5102b07d7d39528b3c84"
              },
              "userId": {
                "type": "string",
                "description": "User's id",
                "example": "654a481000b1c10dd00cd59b"
              },
              "calories": {
                "type": "number",
                "description": "Calories in product",
                "example": 368
              },
              "createdAt": {
                "type": "string",
                "description": "Day summary's creation update date",
                "example": "2023-11-07T15:00:18.935Z"
              },
              "date": {
                "type": "string",
                "description": "Curren date",
                "example": "2023-11-07T15:00:18.935Z"
              },

              "sumId": {
                "type": "string",
                "description": "Day's id",
                "example": "654a481100b1c10dd00cd59d"
              },
              "updateAt": {
                "type": "string",
                "description": "Day summary's creation update date",
                "example": "2023-11-07T18:06:26.474Z"
              },
              "weight": {
                "type": "number",
                "description": "Kcal left to consume",
                "example": 100
              }
              
            }
          },
          "message":{
            "type": "object",
            "description": "Updated summary (or newSummary)",
            "example":"Food deleted"
          },
          "summary": {
            "type": "object",
            "description": "Updated summary (or newSummary)",
            "properties": {
              "id": {
                "type": "string",
                "description": "Sum's id",
                "example": "654a481100b1c10dd00cd59d"
              },
              "date": {
                "type": "string",
                "description": "Day's date",
                "example": "2023-11-07T14:22:09.373Z"
              },
              "userId": {
                "type": "string",
                "description": "User's id",
                "example": "654a481000b1c10dd00cd59b"
              },
              "dailyRate": {
                "type": "number",
                "description": "Daily rate in kcal",
                "example": 1604
              },
              "left": {
                "type": "number",
                "description": "Kcal left to consume",
                "example": 1404
              },
              "consumed": {
                "type": "number",
                "description": "Kcal consumed",
                "example": 200
              },
              
              "percentsOfDailyRate": {
                "type": "number",
                "description": "Percents consumed of daily rate",
                "example": 35.41147132169576
              },
              "createdAt": {
                "type": "date",
                "description": "Day summary's creation date",
                "example": "2023-11-07T14:22:09.374Z"
              },
              "updatedAt": {
                "type": "date",
                "description": "Day summary's update date",
                "example": "2023-11-07T18:06:26.464Z"
              }
              
            }
          }
          
        }
      },
      "GetDayRequest": {
        "type": "object",
        "required": ["date"],
        "properties": {
          "date": {
            "type": "string",
            "description": "Day's date",
            "example": "2023-11-07"
          }
        }
      },
      "GetDayResponse": {
        "type": "object",
        "description": "Day (or newDay) when the product was eaten",
        "properties": {
          "id": {
            "type": "string",
            "description": "Day's id",
            "example": "507f1f77bcf86cd799439011"
          },
          "eatenProducts": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string",
                  "description": "Eaten product's description",
                  "example": "Egg powder omelette"
                },
                "weight": {
                  "type": "number",
                  "description": "Eaten product's weight",
                  "example": 200
                },
                "kcal": {
                  "type": "number",
                  "description": "kcal consumed",
                  "example": 157
                },
                "id": {
                  "type": "string",
                  "description": "Eaten product's id",
                  "example": "9b1deb4d-3b7d-4bad-9bdd-2b0d7b3dcb6d"
                }
              }
            }
          },
          "date": {
            "type": "string",
            "description": "Day's date",
            "example": "2023-11-07"
          },
          "daySummary": {
            "type": "object",
            "description": "Day's summary",
            "properties": {
              "id": {
                "type": "string",
                "description": "Day's id",
                "example": "507f1f77bcf86cd799439011"
              },
              "date": {
                "type": "string",
                "description": "Day's date",
                "example": "2023-11-07"
              },
              "kcalLeft": {
                "type": "number",
                "description": "Kcal left to consume",
                "example": 1194.5
              },
              "kcalConsumed": {
                "type": "number",
                "description": "Kcal consumed",
                "example": 157
              },
              "dailyRate": {
                "type": "number",
                "description": "Daily rate in kcal",
                "example": 1351.5
              },
              "percentsOfDailyRate": {
                "type": "number",
                "description": "Percents consumed of daily rate",
                "example": 11.616722160562338
              },
              "userId": {
                "type": "string",
                "description": "User's id",
                "example": "507f1f77bcf86cd799439012"
              }
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}